+++ src-core/modules/noaa/dsb_deframer.h
@@ -34,6 +34,6 @@ namespace noaa
         // Return state
         int getState();
         // Perform deframing
-        std::vector<std::array<uint8_t, FRAME_SIZE>> work(uint8_t *input, size_t size);
+        std::vector<std::array<uint8_t, FRAME_SIZE>> work(uint8_t *input, int size);
     };
 } // namespace noaa
